refactor(config): move enumerateLeafPaths to a neutral config/leaf-paths module
